Revert "feat(touch): improved frequency analysis calibration (#365)"
This reverts commit 1cbd8d760c9a4a161909227734d9be89a50042a8.

It sadly does not seem like a viable path forward.
