+++ src/cartographer/interfaces/printer.py
@@ -59,7 +59,6 @@ class Endstop(Protocol):
 @dataclass(frozen=True)
 class Sample:
     frequency: float
-    count: int
     time: float
     position: Position | None
     temperature: float
