+++ tests/macros/bed_mesh/test_scan_mesh.py
@@ -136,7 +136,6 @@ class TestBedMeshIntegration:
             sample = Sample(
                 time=float(i * 0.1),
                 frequency=1000.0,
-                count=1,
                 temperature=25.0,
                 position=Position(x=nozzle_x, y=nozzle_y, z=height),
             )
