+++ tests/probes/test_scan_model.py
@@ -31,9 +31,7 @@ def model_factory() -> ScanModelFactory:
 
 
 def test_fit() -> None:
-    samples = [
-        Sample(time=i, count=1, frequency=1 / i, position=Position(0, 0, 0), temperature=0) for i in range(1, 20)
-    ]
+    samples = [Sample(time=i, frequency=1 / i, position=Position(0, 0, 0), temperature=0) for i in range(1, 20)]
 
     fit = ScanModel.fit("test", samples, 0)
